Implement duplicator viewport/render visibility options
This allows a duplicator (as known as dupli parent) to be in a visible
collection so its duplicated objects are visible, however while being
invisible for the final render.

Development note:
Since F12 will only really work once we get a different depsgraph for render
engines and the viewport, I want to be sure that at least viewport is working.
So DEG_iterator_objects_begin needs DAG_EVAL_VIEWPORT evaluation mode.

Not implemented yet:
Integration between this and the particle system emitter visibility. Since we
are doing changes in DNA I think this patch can wait to handle the emitter
visibility as well, to simplify doversioning. That said, the patch can be
reviewed as it is.

Reviewers: sergey

Subscribers: mont29, campbellbarton

Differential Revision: https://developer.blender.org/D2966