Font textbox overflow: Different methods
Sometimes the text doesn't fit. What to do in this case?

* Overflow: The default behaviour still is to overflow the text.
* Truncated: If any text box is defined we can also not draw the text
  that goes outside the text boxes.
* Scale to Fit: For single-text box texts with either width or height defined
  we can scale down the text until it fits.

For now we only support scale to fit for the simplest of the cases.
For proper textbox support we need to iterate over the function
bisecting the scale until we get just the right ammount of lines.

This is not trivial though, and may be better tackled by integrating existing
layout engines such as HarfBuzz.